fully test FileSystem changes
